Remove caching in script
For getting the resulting built binaries, that can be done with a -v argument, while for caching during the installer, the code complexity (particularly in the installer itself) is not worth the potential speedup. The installer is pretty quick anyways, and it is only run during the creation of the docker image.